#31 Rename module and components
