--- src/client/components/MarkdownEditor/MarkdownEditor.spec.js
@@ -1,2 +1,0 @@
-describe('<MarkdownEditor />', () => {
-});
