--- src/client/components/MarkdownEditor/index.js
@@ -1,1 +1,0 @@
-export default require('./MarkdownEditor.react').default;
