+++ src/client/components/MarkdownInput/MarkdownInput.spec.js
@@ -1,0 +1,2 @@
+describe('<MarkdownInput />', () => {
+});
