+++ src/client/components/MarkdownInput/index.js
@@ -1,0 +1,1 @@
+export default require('./MarkdownInput.react').default;
