--- src/client/components/PlainMarkdownEditor/index.js
@@ -1,1 +1,0 @@
-export default require('./PlainMarkdownEditor.react').default;
