+++ src/client/components/PlainMarkdownInput/buttons/BoldButton.js
