+++ src/client/components/PlainMarkdownInput/buttons/HeaderSixButton.js
