+++ src/client/components/PlainMarkdownInput/buttons/ItalicButton.js
