+++ src/client/components/PlainMarkdownInput/buttons/StrikethroughButton.js
