+++ src/client/components/PlainMarkdownInput/buttons/__specs__/BoldButton.spec.js
