+++ src/client/components/PlainMarkdownInput/buttons/__specs__/ItalicButton.spec.js
