+++ src/client/components/PlainMarkdownInput/buttons/__specs__/StrikethroughButton.spec.js
