+++ src/client/components/PlainMarkdownInput/index.js
@@ -1,0 +1,1 @@
+export default require('./PlainMarkdownInput.react').default;
