+++ src/client/components/SlateEditor/SlateEditor.less
@@ -1,4 +1,4 @@
-.react-markdown--plain-markdown-editor__fullscreen-button {
+.react-markdown--plain-markdown-input__fullscreen-button {
   margin-left: auto !important;
 }
