+++ www/index.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>React markdown</title>
+  <title>React Markdown</title>
 </head>
 
 <body style="padding-bottom: 0; overflow-x: hidden;">
